Build Raspberry Pi face ad MVP
