+++ .env.sample
@@ -1,0 +1,14 @@
+APP_ENV=production
+HOST=0.0.0.0
+PORT=8000
+SECRET_KEY=change-me
+CORS_ALLOW_ORIGINS=http://raspberrypi.local:8000
+SQLITE_PATH=./data/app.db
+DB_ECHO=false
+CAMERA_INDEX=0
+FRAME_WIDTH=640
+FRAME_HEIGHT=480
+DLIB_MODEL_DIR=./models
+ID_HASH_SALT=please-change-me
+AD_TEMPLATE_PATH=./templates/ad.html
+PUSH_DELAY_MS=5000
