+++ .github/workflows/deploy.yml
@@ -1,0 +1,50 @@
+name: Deploy to Raspberry Pi
+
+on:
+  push:
+    branches: [ "main" ]
+
+jobs:
+  deploy:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+
+      - name: Create .env from sample
+        run: |
+          mkdir -p data models templates static
+          cp .env.sample .env
+
+      - name: Add SSH key
+        uses: webfactory/ssh-agent@v0.9.0
+        with:
+          ssh-private-key: |
+            ${{ secrets.PI_SSH_KEY }}
+
+      - name: Copy files to Pi
+        run: |
+          rsync -avz --delete \
+            -e "ssh -o StrictHostKeyChecking=no" \
+            ./ ${{ secrets.PI_USER }}@${{ secrets.PI_HOST }}:${{ vars.PI_APP_DIR }}/
+
+      - name: Remote install & restart
+        run: |
+          ssh -o StrictHostKeyChecking=no ${{ secrets.PI_USER }}@${{ secrets.PI_HOST }} << 'EOF'
+          set -e
+          cd ${{ vars.PI_APP_DIR }}
+          python3 -m venv .venv
+          . .venv/bin/activate
+          pip install --upgrade pip
+          pip install -r requirements.txt
+          pkill -f "python.*app.py" || true
+          pkill -f "python.*camera_loop.py" || true
+          nohup .venv/bin/python camera_loop.py > cam.log 2>&1 &
+          nohup .venv/bin/python app.py > web.log 2>&1 &
+          echo "Deployed and restarted."
+          EOF
