+++ .gitignore
@@ -1,0 +1,8 @@
+.venv/
+__pycache__/
+*.py[cod]
+*.log
+.env
+cam.log
+web.log
+data/*.db
