+++ templates/ad.html
@@ -1,0 +1,32 @@
+<!doctype html>
+<html>
+<head>
+  <meta charset="utf-8"/>
+  <meta name="viewport" content="width=device-width,initial-scale=1"/>
+  <title>Smart Ad</title>
+  <style>
+    body { font-family: system-ui, sans-serif; margin: 0; background:#0b1220; color:#e6edf3; }
+    .wrap { display:grid; place-items:center; min-height:100vh; }
+    .card { background:#111b2a; padding:32px; border-radius:20px; box-shadow:0 8px 30px rgba(0,0,0,.35); width:min(720px, 92vw); }
+    h1 { margin: 0 0 12px; }
+    .muted { opacity:.7; }
+    .pill { display:inline-block; padding:6px 10px; border-radius:999px; background:#1f2b41; font-size:12px; }
+    ul { line-height:1.8; }
+  </style>
+</head>
+<body>
+  <div class="wrap">
+    <div class="card">
+      <div class="pill">匿名會員：{{ member_id }}</div>
+      <h1>{{ headline }}</h1>
+      <p class="muted">特別推薦：</p>
+      <ul>
+        {% for x in purchases %}
+        <li>{{ x['sku'] }} — 上次 {{ x['ago'] }} 前購買。<strong>今日{{ x['offer'] }}</strong></li>
+        {% endfor %}
+      </ul>
+      <p class="muted">* 僅使用不可逆的匿名 ID；不會上傳或保存影像。</p>
+    </div>
+  </div>
+</body>
+</html>
